Separated the logic of handling the nodes and the algorithm
The code that is relevant to the algorithm is saved under
the fit_algorithm class and all the information relevant to the nodes
is in the nodes_manager. This is done since we wish to separate the
responsibilities between both classes.
The algorithm might have a state that is updated through each of its
steps so its necessary that it has some kind of instance.